requirements.txt
flask==2.3.3
python-telegram-bot==20.7
google-genai==1.2.0
Pillow==10.1.0